refactor pruner to use strategy pattern
