+++ serialization/src/main/java/org/finos/rune/mapper/processor/EmptyAttributePruningStrategy.java
@@ -1,0 +1,10 @@
+package org.finos.rune.mapper.processor;
+
+import com.rosetta.model.lib.RosettaModelObjectBuilder;
+
+public class EmptyAttributePruningStrategy implements PruningStrategy {
+    @Override
+    public void prune(RosettaModelObjectBuilder builder) {
+        builder.prune();
+    }
+}
