+++ serialization/src/main/java/org/finos/rune/mapper/processor/PreSerializationPruner.java
@@ -1,30 +1,27 @@
 package org.finos.rune.mapper.processor;
 
-import com.rosetta.model.lib.GlobalKey;
 import com.rosetta.model.lib.RosettaModelObject;
 import com.rosetta.model.lib.RosettaModelObjectBuilder;
-import com.rosetta.model.lib.meta.FieldWithMeta;
-import com.rosetta.model.lib.meta.GlobalKeyFields;
 import com.rosetta.model.lib.path.RosettaPath;
 import com.rosetta.model.lib.process.AttributeMeta;
 import com.rosetta.model.lib.process.BuilderProcessor;
 
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 
 public class PreSerializationPruner implements BuilderProcessor {
-    private final Set<GlobalReferenceRecord> globalReferences;
+    private final List<PruningStrategy> pruningStrategies;
 
-    public PreSerializationPruner(Set<GlobalReferenceRecord> globalReferences) {
-        this.globalReferences = globalReferences;
+    public PreSerializationPruner(List<PruningStrategy> pruningStrategies) {
+        this.pruningStrategies = pruningStrategies;
     }
 
     @Override
     public <R extends RosettaModelObject> boolean processRosetta(RosettaPath path, Class<R> rosettaType, RosettaModelObjectBuilder builder, RosettaModelObjectBuilder parent, AttributeMeta... metas) {
         if (builder != null) {
-            pruneGlobalKeys(builder);
-            pruneEmptyAttributes(builder);
+            for (PruningStrategy pruningStrategy : pruningStrategies) {
+                pruningStrategy.prune(builder);
+            }
             return true;
         }
         return false;
@@ -63,28 +60,4 @@ public class PreSerializationPruner implements BuilderProcessor {
         throw new UnsupportedOperationException("PreSerializationPruner report not supported");
     }
 
-    private void pruneGlobalKeys(RosettaModelObjectBuilder builder) {
-        if (builder instanceof GlobalKey.GlobalKeyBuilder) {
-            GlobalKey.GlobalKeyBuilder globalKeyBuilder = (GlobalKey.GlobalKeyBuilder) builder;
-            GlobalKeyFields.GlobalKeyFieldsBuilder globalKeyFields = globalKeyBuilder.getMeta();
-            String globalKey = globalKeyFields.getGlobalKey();
-            if (globalKey != null) {
-                GlobalReferenceRecord globalReferenceRecord = new GlobalReferenceRecord(getType(builder), globalKey);
-                if (!globalReferences.contains(globalReferenceRecord)) {
-                    globalKeyFields.setGlobalKey(null);
-                }
-            }
-        }
-    }
-
-    private Class<?> getType(RosettaModelObjectBuilder builder) {
-        if (builder instanceof FieldWithMeta.FieldWithMetaBuilder) {
-            return ((FieldWithMeta.FieldWithMetaBuilder<?>)builder).getValueType();
-        }
-        return builder.getType();
-    }
-
-    private void pruneEmptyAttributes(RosettaModelObjectBuilder builder) {
-        builder.prune();
-    }
 }
