+++ serialization/src/main/java/org/finos/rune/mapper/processor/PruningStrategy.java
@@ -1,0 +1,7 @@
+package org.finos.rune.mapper.processor;
+
+import com.rosetta.model.lib.RosettaModelObjectBuilder;
+
+public interface PruningStrategy {
+    void prune(RosettaModelObjectBuilder builder);
+}
