+++ serialization/src/main/java/org/finos/rune/mapper/processor/SerializationPreProcessor.java
@@ -1,9 +1,12 @@
 package org.finos.rune.mapper.processor;
 
+import com.google.common.collect.Lists;
 import com.rosetta.model.lib.RosettaModelObject;
 import com.rosetta.model.lib.RosettaModelObjectBuilder;
 import com.rosetta.model.lib.path.RosettaPath;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 public class SerializationPreProcessor {
@@ -12,9 +15,13 @@ public class SerializationPreProcessor {
         RosettaPath path = RosettaPath.valueOf(rosettaModelObject.getType().getSimpleName());
         GlobalReferenceCollector globalReferenceCollector = new GlobalReferenceCollector();
         rosettaModelObject.process(path, globalReferenceCollector);
-
         Set<GlobalReferenceRecord> globalReferences = globalReferenceCollector.getGlobalReferences();
-        PreSerializationPruner preSerializationPruner = new PreSerializationPruner(globalReferences);
+
+        GlobalKeyPruningStrategy globalKeyPruningStrategy = new GlobalKeyPruningStrategy(globalReferences);
+        EmptyAttributePruningStrategy emptyAttributePruningStrategy = new EmptyAttributePruningStrategy();
+        List<PruningStrategy> pruningStrategyList = Lists.newArrayList(globalKeyPruningStrategy, emptyAttributePruningStrategy);
+
+        PreSerializationPruner preSerializationPruner = new PreSerializationPruner(pruningStrategyList);
         RosettaModelObjectBuilder builder = rosettaModelObject.toBuilder();
         builder.process(path, preSerializationPruner);
         return buildAndCast(builder);
